refactor: reorganise webview implementations
