--- src/Packages/Passport/Runtime/Scripts/Private/UI/AndroidPassportWebView.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ca562827a0f98a7489504f4aa7ffe2dc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
