+++ src/Packages/Passport/Runtime/Scripts/Private/UI/WebViews/iOSPassportWebView.cs
@@ -6,22 +6,22 @@ using UnityEngine;
 using UnityEngine.UI;
 using Immutable.Passport.Core.Logging;
 
-#if UNITY_ANDROID && !UNITY_EDITOR
+#if UNITY_IOS && !UNITY_EDITOR
 using Vuplex.WebView;
 #endif
 
 namespace Immutable.Passport
 {
     /// <summary>
-    /// Android implementation of IPassportWebView using Vuplex WebView
-    /// Provides embedded WebView functionality within the Unity app (not external browser)
-    /// This is different from AndroidPassportWebView which uses external browser for auth flows
+    /// iOS implementation of IPassportWebView using Vuplex WebView
+    /// Provides embedded WebView functionality within the Unity app
+    /// Consistent with Android and macOS Vuplex implementations
     /// </summary>
-    public class AndroidVuplexWebView : IPassportWebView
+    public class iOSPassportWebView : IPassportWebView
     {
-        private const string TAG = "[AndroidVuplexWebView]";
+        private const string TAG = "[iOSPassportWebView]";
 
-#if UNITY_ANDROID && !UNITY_EDITOR
+#if UNITY_IOS && !UNITY_EDITOR
         private CanvasWebViewPrefab? _webViewPrefab;
 #endif
         private readonly Dictionary<string, Action<string>> _jsHandlers = new Dictionary<string, Action<string>>();
@@ -33,7 +33,7 @@ namespace Immutable.Passport
         public event Action? OnLoadStarted;
 
         // Safe access - check initialization
-#if UNITY_ANDROID && !UNITY_EDITOR
+#if UNITY_IOS && !UNITY_EDITOR
         public bool IsVisible => _webViewPrefab?.Visible ?? false;
         public string CurrentUrl => _webViewPrefab?.WebView?.Url ?? "";
 #else
@@ -41,7 +41,7 @@ namespace Immutable.Passport
         public string CurrentUrl => "";
 #endif
 
-        public AndroidVuplexWebView(RawImage canvasReference)
+        public iOSPassportWebView(RawImage canvasReference)
         {
             _canvasReference = canvasReference ?? throw new ArgumentNullException(nameof(canvasReference));
         }
@@ -54,10 +54,10 @@ namespace Immutable.Passport
                 return;
             }
 
-#if UNITY_ANDROID && !UNITY_EDITOR
+#if UNITY_IOS && !UNITY_EDITOR
             try
             {
-                PassportLogger.Info($"{TAG} Initializing Vuplex WebView...");
+                PassportLogger.Info($"{TAG} Initializing iOS WebView...");
 
                 // Start async initialization but don't wait
                 InitializeAsync(config).Forget();
@@ -68,28 +68,38 @@ namespace Immutable.Passport
                 throw;
             }
 #else
-            PassportLogger.Warn($"{TAG} Vuplex WebView is only supported on Android builds, not in editor");
+            PassportLogger.Warn($"{TAG} Vuplex WebView is only supported on iOS builds, not in editor");
             _isInitialized = true;
 #endif
         }
 
-#if UNITY_ANDROID && !UNITY_EDITOR
+#if UNITY_IOS && !UNITY_EDITOR
         private async UniTaskVoid InitializeAsync(PassportWebViewConfig config)
         {
             try
             {
                 // Create WebView prefab and parent to Canvas
                 _webViewPrefab = CanvasWebViewPrefab.Instantiate();
-                _webViewPrefab.Native2DModeEnabled = true;
+                _webViewPrefab.Native2DModeEnabled = false; // Disable Native2DMode to avoid Unity integration issues
+
+                // Set reasonable resolution - much lower to avoid texture size issues
+                _webViewPrefab.Resolution = 1.0f; // 1px per Unity unit - creates 800x600px texture
 
                 // Must be child of Canvas for Vuplex to work
                 _webViewPrefab.transform.SetParent(_canvasReference.canvas.transform, false);
 
-                // Set up full-screen layout for Native 2D Mode
+                // Set WebView size based on configuration
                 var rect = _webViewPrefab.GetComponent<RectTransform>();
-                rect.anchorMin = Vector2.zero;
-                rect.anchorMax = Vector2.one;
-                rect.offsetMin = rect.offsetMax = Vector2.zero;
+                rect.anchorMin = new Vector2(0.5f, 0.5f); // Center anchor
+                rect.anchorMax = new Vector2(0.5f, 0.5f);
+
+                // Use configured dimensions or fallback to reasonable defaults
+                float width = config.Width > 0 ? config.Width : 1000;
+                float height = config.Height > 0 ? config.Height : 700;
+                rect.sizeDelta = new Vector2(width, height);
+                rect.anchoredPosition = Vector2.zero; // Center position
+
+                PassportLogger.Info($"{TAG} Using WebView dimensions: {width}x{height}");
 
                 // Wait for WebView initialization
                 await _webViewPrefab.WaitUntilInitialized();
@@ -131,11 +141,11 @@ namespace Immutable.Passport
                 _webViewPrefab.WebView.LoadFailed += (sender, failedArgs) => PassportLogger.Warn($"{TAG} Load failed: {failedArgs.NativeErrorCode} for {failedArgs.Url}");
 
                 _isInitialized = true;
-                PassportLogger.Info($"{TAG} Vuplex WebView initialized successfully");
+                PassportLogger.Info($"{TAG} iOS WebView initialized successfully");
             }
             catch (Exception ex)
             {
-                PassportLogger.Error($"{TAG} Failed to initialize async: {ex.Message}");
+                PassportLogger.Error($"{TAG} Failed to initialize iOS WebView: {ex.Message}");
                 throw;
             }
         }
@@ -143,22 +153,22 @@ namespace Immutable.Passport
 
         public void LoadUrl(string url)
         {
-#if UNITY_ANDROID && !UNITY_EDITOR
+#if UNITY_IOS && !UNITY_EDITOR
             if (!_isInitialized || _webViewPrefab?.WebView == null)
             {
-                PassportLogger.Error($"{TAG} Cannot load URL - WebView not initialized");
+                PassportLogger.Error($"{TAG} Cannot load URL - iOS WebView not initialized");
                 return;
             }
 
             _webViewPrefab.WebView.LoadUrl(url);
 #else
-            PassportLogger.Warn($"{TAG} LoadUrl not supported in editor mode");
+            PassportLogger.Warn($"{TAG} LoadUrl not supported in iOS editor mode");
 #endif
         }
 
         public void Show()
         {
-#if UNITY_ANDROID && !UNITY_EDITOR
+#if UNITY_IOS && !UNITY_EDITOR
             if (_webViewPrefab != null)
             {
                 _webViewPrefab.Visible = true;
@@ -168,7 +178,7 @@ namespace Immutable.Passport
 
         public void Hide()
         {
-#if UNITY_ANDROID && !UNITY_EDITOR
+#if UNITY_IOS && !UNITY_EDITOR
             if (_webViewPrefab != null)
             {
                 _webViewPrefab.Visible = false;
@@ -178,10 +188,10 @@ namespace Immutable.Passport
 
         public void ExecuteJavaScript(string js)
         {
-#if UNITY_ANDROID && !UNITY_EDITOR
+#if UNITY_IOS && !UNITY_EDITOR
             if (!_isInitialized || _webViewPrefab?.WebView == null)
             {
-                PassportLogger.Error($"{TAG} Cannot execute JavaScript - WebView not initialized");
+                PassportLogger.Error($"{TAG} Cannot execute JavaScript - iOS WebView not initialized");
                 return;
             }
 
@@ -200,7 +210,7 @@ namespace Immutable.Passport
 
         public void Dispose()
         {
-#if UNITY_ANDROID && !UNITY_EDITOR
+#if UNITY_IOS && !UNITY_EDITOR
             if (_webViewPrefab != null)
             {
                 _webViewPrefab.Destroy();
